+++ metricservices/client/standard.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 	"net/url"
 	"strings"
-	"time"
 
 	"github.com/tidepool-org/platform/errors"
 	"github.com/tidepool-org/platform/log"
@@ -33,13 +32,12 @@ func NewStandard(versionReporter version.Reporter, name string, config *Config) 
 		return nil, errors.New("client", "config is missing")
 	}
 
-	config = config.Clone()
 	if err := config.Validate(); err != nil {
 		return nil, errors.Wrap(err, "client", "config is invalid")
 	}
 
 	httpClient := &http.Client{
-		Timeout: time.Duration(config.RequestTimeout) * time.Second,
+		Timeout: config.Timeout,
 	}
 
 	return &Standard{