+++ service/server/config.go
@@ -2,29 +2,62 @@ package server
 
 import (
 	"os"
+	"strconv"
+	"time"
 
+	"github.com/tidepool-org/platform/config"
 	"github.com/tidepool-org/platform/errors"
 )
 
 type Config struct {
-	Address string `json:"address"`
-	TLS     *TLS   `json:"tls"`
-	Timeout int    `json:"timeout" default:"60"`
+	Address            string
+	TLS                bool
+	TLSCertificateFile string
+	TLSKeyFile         string
+	Timeout            time.Duration
 }
 
-type TLS struct {
-	CertificateFile string `json:"certificateFile"`
-	KeyFile         string `json:"keyFile"`
+func NewConfig() *Config {
+	return &Config{
+		TLS:     true,
+		Timeout: 60 * time.Second,
+	}
+}
+
+func (c *Config) Load(configReporter config.Reporter) error {
+	if configReporter == nil {
+		return errors.New("server", "config reporter is missing")
+	}
+
+	c.Address = configReporter.StringOrDefault("address", "")
+	if tlsString, found := configReporter.String("tls"); found {
+		tls, err := strconv.ParseBool(tlsString)
+		if err != nil {
+			return errors.New("server", "tls is invalid")
+		}
+		c.TLS = tls
+	}
+	c.TLSCertificateFile = configReporter.StringOrDefault("tls_certificate_file", "")
+	c.TLSKeyFile = configReporter.StringOrDefault("tls_key_file", "")
+	if timeoutString, found := configReporter.String("timeout"); found {
+		timeout, err := strconv.ParseInt(timeoutString, 10, 0)
+		if err != nil {
+			return errors.New("server", "timeout is invalid")
+		}
+		c.Timeout = time.Duration(timeout) * time.Second
+	}
+
+	return nil
 }
 
 func (c *Config) Validate() error {
 	if c.Address == "" {
 		return errors.New("server", "address is missing")
 	}
-	if c.TLS != nil {
-		if c.TLS.CertificateFile == "" {
+	if c.TLS {
+		if c.TLSCertificateFile == "" {
 			return errors.New("server", "tls certificate file is missing")
-		} else if fileInfo, err := os.Stat(c.TLS.CertificateFile); err != nil {
+		} else if fileInfo, err := os.Stat(c.TLSCertificateFile); err != nil {
 			if !os.IsNotExist(err) {
 				return errors.Wrap(err, "server", "unable to stat tls certificate file")
 			}
@@ -32,9 +65,9 @@ func (c *Config) Validate() error {
 		} else if fileInfo.IsDir() {
 			return errors.New("server", "tls certificate file is a directory")
 		}
-		if c.TLS.KeyFile == "" {
+		if c.TLSKeyFile == "" {
 			return errors.New("server", "tls key file is missing")
-		} else if fileInfo, err := os.Stat(c.TLS.KeyFile); err != nil {
+		} else if fileInfo, err := os.Stat(c.TLSKeyFile); err != nil {
 			if !os.IsNotExist(err) {
 				return errors.Wrap(err, "server", "unable to stat tls key file")
 			}
@@ -43,22 +76,19 @@ func (c *Config) Validate() error {
 			return errors.New("server", "tls key file is a directory")
 		}
 	}
-	if c.Timeout < 0 {
+	if c.Timeout <= 0 {
 		return errors.New("server", "timeout is invalid")
 	}
+
 	return nil
 }
 
 func (c *Config) Clone() *Config {
-	clone := &Config{
-		Address: c.Address,
-		Timeout: c.Timeout,
+	return &Config{
+		Address:            c.Address,
+		TLS:                c.TLS,
+		TLSCertificateFile: c.TLSCertificateFile,
+		TLSKeyFile:         c.TLSKeyFile,
+		Timeout:            c.Timeout,
 	}
-	if c.TLS != nil {
-		clone.TLS = &TLS{
-			CertificateFile: c.TLS.CertificateFile,
-			KeyFile:         c.TLS.KeyFile,
-		}
-	}
-	return clone
 }