+++ tools/migrate_sessions_expand/migrate_sessions_expand.go
@@ -30,7 +30,7 @@ const (
 	VerboseFlag   = "verbose"
 	DryRunFlag    = "dry-run"
 	AddressesFlag = "addresses"
-	SSLFlag       = "ssl"
+	TLSFlag       = "tls"
 	SecretFlag    = "secret"
 )
 
@@ -82,8 +82,8 @@ func initializeApplication() (*cli.App, error) {
 			Usage: "comma-delimited list of address(es) to mongo database (host:port)",
 		},
 		cli.BoolFlag{
-			Name:  fmt.Sprintf("%s,%s", SSLFlag, "s"),
-			Usage: "use SSL to connect to mongo database",
+			Name:  fmt.Sprintf("%s,%s", TLSFlag, "t"),
+			Usage: "use TLS to connect to mongo database",
 		},
 		cli.StringFlag{
 			Name:   SecretFlag,
@@ -140,24 +140,18 @@ func executeApplication(versionReporter version.Reporter, context *cli.Context) 
 
 func buildConfigFromContext(context *cli.Context) (*Config, error) {
 	config := &Config{
-		Log: &log.Config{
-			Level: "info",
-		},
-		Mongo: &mongo.Config{
-			Timeout: app.DurationAsPointer(60 * time.Second),
-		},
+		Log:   log.NewConfig(),
+		Mongo: mongo.NewConfig(),
 	}
 
 	if context.Bool(VerboseFlag) {
 		config.Log.Level = "debug"
+	} else {
+		config.Log.Level = "info"
 	}
-	config.Mongo.Addresses = context.String(AddressesFlag)
-	if context.Bool(SSLFlag) {
-		config.Mongo.SSL = true
-	}
-	if context.Bool(DryRunFlag) {
-		config.DryRun = true
-	}
+	config.Mongo.Addresses = app.SplitStringAndRemoveWhitespace(context.String(AddressesFlag), ",")
+	config.Mongo.TLS = context.Bool(TLSFlag)
+	config.DryRun = context.Bool(DryRunFlag)
 	config.Secret = context.String(SecretFlag)
 	if config.Secret == "" {
 		return nil, errors.New("migrate_sessions_expand", "secret is missing")